feat: change DB to redis to verify email
